niffler-e-2-e-tests-python/requirements.txt
playwright
pytest-playwright
pytest
python-dotenv
faker